modify Networking
DataManager, SessionManager, Networking으로 책임 분리
